feat: merge

+++ src/main/java/com/newBig/system/infrastructure/persistence/FuncionarioRepo.java
@@ -1,5 +1,6 @@
 package com.newBig.system.infrastructure.persistence;
 
+import com.newBig.system.Main;
 import com.newBig.system.domain.model.Cliente;
 import com.newBig.system.domain.model.Funcionario;
 import jakarta.persistence.EntityManager;
@@ -28,9 +29,15 @@ public class FuncionarioRepo {
 
     public void update(Funcionario dados) {
         try{
-            em.getTransaction().begin();
-            em.merge(dados);
-            em.getTransaction().commit();
+            if(dados == null){
+                System.out.println("funcionario nâo encontrado no banco");
+                Main.main(null);
+            }
+            else{
+                em.getTransaction().begin();
+                em.merge(dados);
+                em.getTransaction().commit();
+            }
         }
         catch (Exception e){
             em.getTransaction().rollback(); /* Não deixa salvar se dar erro */
@@ -40,9 +47,21 @@ public class FuncionarioRepo {
     }
 
     public void delete(Funcionario dados) {
-        em.getTransaction().begin();
-        em.remove(em.contains(dados) ? dados : em.merge(dados));
-        em.getTransaction().commit();
+        try{
+            if(dados == null){
+                System.out.println("funcionario nâo encontrado no banco");
+                Main.main(null);
+            }
+            else{
+                em.getTransaction().begin();
+                em.remove(em.contains(dados) ? dados : em.merge(dados));
+                em.getTransaction().commit();
+            }
+        } catch (Exception e) {
+            em.getTransaction().rollback(); /* Não deixa salvar se dar erro */
+            System.out.println("Erro ao salvar!! Nada foi salvo no Banco de dados");
+        }
+
     }
 
     public List<Funcionario> BuscaFuncionarios() {
@@ -67,5 +86,11 @@ public class FuncionarioRepo {
                 )
                 .setParameter("nome", "%" + nome + "%") /*Define o valor do parametro :nome*/
                 .getResultList(); /*Retorna a lista*/
+    }
+
+    public boolean consultaId(Long id){
+        return em.createQuery(
+                "SELECT COUNT(e) > 0 FROM Funcionario e WHERE e.id = :id", Boolean.class
+        ).setParameter("id", id).getSingleResult();
     }
 }
